backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
PyJWT
python-dotenv
stripe
Pillow

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
